add reusable card component to Games
