+++ app/src/components/Card/styles.module.scss
@@ -1,0 +1,67 @@
+li {
+  list-style: none;
+}
+
+.button {
+  margin-left: 1rem;
+}
+
+.grid {
+  display: grid;
+  grid-gap: 1rem;
+}
+
+@supports (width: min(30%, 100%)) {
+  .grid {
+    grid-template-columns: repeat(auto-fit, minmax(min(30%, 100%), 1fr));
+  }
+}
+
+// for cards
+
+.dropshadow {
+  filter: drop-shadow(1px 2px 4px hsl(210deg, 2%, 58%));
+}
+
+.box {
+  width: 150px;
+  height: 150px;
+  background-color: white;
+  border-radius: 8px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+}
+
+.wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.cardthumb {
+  display: flex;
+  width: 30%;
+  height: auto;
+}
+
+.detailsimage {
+  display: flex;
+  width: 50%;
+}
+
+.container {
+  display: flex;
+}
+
+.left {
+  flex: 1;
+  padding: 1rem;
+}
+
+.right {
+  flex: 1;
+  padding: 1rem;
+}
+.description {
+}
